Added noToast field for "cmd" extension messages
If eventMsg.data contains ".noToast" with a true value and
the command results in an error, the error message toast will
be suppressed.  That lets the extension handle errors internally
without the user seeing a possibly-confusing message.

One use case is when the extension needs to issue a command
that might not exist in older versions of the controller
firmware, handling the error quietly for backwards compatibility.